+++ config-overrides.js
@@ -1,12 +1,12 @@
 /* config-overrides.js | react-app-rewired */
 
-const HtmlWebpackPlugin = require('html-webpack-plugin');
-const fs = require('fs');
-const path = require('path');
+import HtmlWebpackPlugin from 'html-webpack-plugin';
+import fs from 'fs';
+import path from 'path';
+
 const appLocalDomainName = process.env.APP_LOCAL_DOMAIN_NAME;
 
 process.env.REACT_APP_API_VERSION = process.env.REACT_APP_API_VERSION || process.env.API_VERSION || 'v1';
-process.env.REACT_APP_API_KEYS_PREFIX = process.env.REACT_APP_API_KEYS_PREFIX || process.env.API_KEYS_PREFIX || 'sk-gsu-';
 
 /*
 https://github.com/timarney/react-app-rewired
@@ -23,9 +23,21 @@ npm install --save-dev \
    stream-browserify \
    vm-browserify \
    tty-browserify
+
+To avoid issues with the app URL and execution:
+
+File: "package.json"
+
+- If there is a "type": "module" attribute, rename it to e.g. "type1"
+- If there is a "homepage": "..." attribute, rename it to e.g. "homepage1" (unless you really need a URL # suffix)
+- If there is a "eslintConfig": "..." attribute, rename it to e.g. "eslintConfig1"
+- Add the following entry to "scripts":
+    "start-dev:react-app-rewired": "bash ../node_modules/genericsuite-fe-scripts/scripts/change_env_be_endpoint.sh dev && npx react-app-rewired start",
+- Install typescript: "npm installl -D typescript" or "npm installl -D -w ui typescript"
+- After changing "package.json", run "npm update" or "npm update -w ui" before start the app.
 */
 
-module.exports = {
+export default {
   // The Webpack config to use when compiling your react app for development or production.
   webpack: function (config, env) {
     // ...add your webpack config
@@ -59,6 +71,7 @@ module.exports = {
       // Create the default config by calling configFunction with the proxy/allowedHost parameters
       const config = configFunction(proxy, allowedHost);
       config.allowedHosts = [appLocalDomainName, "localhost", "127.0.0.1"]; // To avoid "Invalid Host header" error
+      config.hostName = appLocalDomainName;
       config.historyApiFallback = true;
       config.hot = true;
       config.compress = true;
@@ -71,7 +84,7 @@ module.exports = {
         // set the file paths & passphrase.
         config.https = {
           key: fs.readFileSync(path.resolve(__dirname, `${appLocalDomainName}.key`), 'utf8'),
-          cert: fs.readFileSync(path.resolve(__dirname, `${appLocalDomainName}.crt`), 'utf8'),
+          cert: fs.readFileSync(path.resolve(__dirname, `${appLocalDomainName}.chain.crt`), 'utf8'),
           ca: fs.readFileSync(path.resolve(__dirname, 'ca.crt'), 'utf8'),
           passphrase: process.env.REACT_HTTPS_PASS
         };
@@ -102,12 +115,12 @@ module.exports = {
     config.fallback = {
       "fs": false,
       // Uncomment as needed (see the npm install note above):
-      // "os": require.resolve("os-browserify/browser"),
-      // "url": require.resolve("url"),
-      // "crypto": require.resolve("crypto-browserify"),
-      // "stream": require.resolve("stream-browserify"),
-      // "vm": require.resolve("vm-browserify"),
-      // "tty": require.resolve("tty-browserify"),
+      "os": require.resolve("os-browserify/browser"),
+      "url": require.resolve("url"),
+      "crypto": require.resolve("crypto-browserify"),
+      "stream": require.resolve("stream-browserify"),
+      "vm": require.resolve("vm-browserify"),
+      "tty": require.resolve("tty-browserify"),
     };
     return config;
   },